renamed and added measurement script parts;moved current to experimental.
