+++ src/measure_cross_sections_part3.py
@@ -1,0 +1,15 @@
+'''
+Created on 9 Dec 2012
+
+@author: kreczko
+
+Episode 3 - Unfolding
+- Read histograms from Episode 2
+- unfold the fit results for central result and systematics
+- combine results for unfolded and not unfolded fit results
+- write everything* to JSON and ROOT files
+
+* all channels + combination, central + systematic results
+'''
+
+
